// src/types.ts
export interface TDMeta {
  isDarkMode: boolean
}
// ...
export enum AlignStyle {
  Start = 'start',
  Middle = 'middle',
  End = 'end',
  Justify = 'justify',
}